improve exception message
